backend/pyproject.toml
[project]
name = "sisy-backend"
version = "1.0.0"
description = "SiSy Backend - AI chat with FastAPI, LangChain, and Opik"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "opik>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[tool.uv]
dev-dependencies = []

# Not building as a package, just using for dependency management
